Performance optimizations.
Cell drawing is now O(n) instead of O(n^2). However, as a result,
clearing the cells is now in O(n^2). This still makes the average
experience far smoother, and it allows for a higher grid size (moved up
to 1,024 * 1,024 from 256 * 256). When updating to the next state,
however, it slows down a fair amount. This breaks autoplay for higher
grid sizes (16,384 * 16,384), so it will have to be rectified at a later
time.